new file for motor things, moved power function into sensors
