+++ sensor.go
@@ -60,3 +60,31 @@ func (s *irSensor) getButton() int {
 	value, _ := strconv.Atoi(v)
 	return value
 }
+
+func getPower() (float64, float64, float64, float64) {
+	p := ev3dev.PowerSupply("")
+	p = ev3dev.PowerSupply(p.String()) // Cache the driver name if not given.
+
+	v, err := p.Voltage()
+	if err != nil {
+		log.Fatalf("could not read voltage: %v", err)
+	}
+
+	i, err := p.Current()
+	if err != nil {
+		log.Fatalf("could not read current: %v", err)
+	}
+
+	vMax, err := p.VoltageMax()
+	if err != nil {
+		log.Fatalf("could not read max design voltage: %v", err)
+	}
+
+	vMin, err := p.VoltageMin()
+	if err != nil {
+		log.Fatalf("could not read min design voltage: %v", err)
+	}
+
+	log.Printf("current power statW: V=%.2fV I=%.0fmA P=%.3fW (designed voltage range:%.2fV-%.2fV)\n", v, i, i*v/1000, vMin/10, vMax/10)
+	return v, i, vMax, vMin
+}
